Moved HTTP handling to requests
We can now have a timeout and better separation of concerns. Feedparser is now
only a parser. Mocking is enforced in tests, they're faster now even though
there is still room for improvement.

Added a model handling parsing-related attributes for feeds (etags,
subscribers) that should eventually be merged with the favicon model.

Updatefeeds now just stuffs jobs in the queue, it's no longer synchronous and
checks that the feed wasn't updated to little time ago.